Creates a "hello world," which displays a random image.
The cellular automata state is represtented as a 32-bit integer, but for
ease of visualization I use only the lower 24 bits and convert those to
a full color.

For the game of life, the next state (live/dead) depends only on the
current state and the sum of the 8 neighbors. My plan is to have many
parallel games of life running, one for each 3 bits, so 8 games running
in parallel if using the full color range.

To demonstrate only a single game of life, I also generate a starting
state using only a single green bit. This can use the same code path but
will be easier to see what is going on.

Note: I recognize that this is more complicated than it really needs to
be, but I thought it could be a fun art project to see what happens
using the full 24 bits color.

[Edit] Rebased to remove build output from repository. Also, added
documentation for public methods.

[Edit2] Renames Cell to Field. Also, I change the validation to panic
on bad width/height values instead of return errors. This makes the code
more consistent with how the image package works.

Change-Id: Ic65903079d62d82f9521142f1d989880ece52504