Plays Conway's Game of Life once per second.
The new file cycle.go implements Conway's Game of Life using some
bitwise arithmetic. This means there are actually many games being
played in parallel.

I realized after viewing it http://23.251.159.233/
that only the bright green bit is very visible. Generating an initial
state using only this bit is unnecessary, so I remove the
RandomGreenCell method.

Also, I was getting annoyed by having to check for errors every time I
call NewCell with valid inputs, so I removed the error from the
function. This seems consistent with the functions in the image package
where it is possible that the Rectangle is not well formed, but none of
the New* functions return an error.

The cell.Run method returns a channel rather than the simpler method of
just having a NextState function to allow updates to happen
asynchronously in the background. It will also allow me to distribute
computation to multiple semi-independent nodes, where it wouldn't make
sense to call an RPC on each to go to the next state.

Change-Id: I7bd347e9659cb7bab7302eaeb2b72246e6bd60b8