Push notifications of new Game of Life states with Websockets.
This also caches the last 2 minutes of game of life state. This way, I
can notify the client of the new image URL and swap the old URL out
without worrying much about browser caching.

Alternatives are to send a Data URI or somehow force the image to
refresh (and not use the cached version), so this seems like the easiest
way to push updates. Caching 2 minutes of history has the additional
benefit that if I like some animation, I can download the images within
2 minutes of seeing it.

Change-Id: I9ca18745a2b7b0480f210272302fec976316a4de