Separates Game of Life into pub & web services.
The pub service is responsible for computing the state for the Game of
Life and communicating it to the web server listeners.

The web service is responsible for converting the game of life state
into an image and displaying the state to browser clients.

Since there isn't a real need for reliability, the pub/web services communicate
over ZeroMQ Pub/Sub sockets. Losing a state update doesn't matter too
much since the server should be sending another update within a second.

To speed up builds, I followed the Docker best practice of layering the
Docker images. Now we do not need to build ZeroMQ every time we change
something in our package.

Change-Id: Ib167a394be69014f50ac96c76b46abacc9e7e666