New package for commands.
Now the artist command is called artgen. I plan to include more
commands, such as artrate, which will be the critiquing code.

Deletes the unused code for moving in lines. I think trying to train
from timelapses will be a lot harder if restricted to certain
strategies. Human-generated timelapses would only have a sequence of
actions.