Modify rating formula.
Adds tests for ratings, too.

Now colors at the endpoints get a rating of 0. Blue does not like images
with no PICO8_BLUE, but he also does not like images that are 100% that
color, either.

This drops the average rating to about 12% from 60% before, but it gives
us a wider range of opinions to work with.