Can now paint on right side of screen.
Adds a test for painting to the right on the right side of the screen.
There was a bug in the maximum X coordinate used where it would not
consider pixels on the right half of the screen.

Also, improve the explanation view in the debug mode to help debug where
it was getting stuck.