Allow switching rating function in strategies.
This lets us test simulation function whose behavior depends on having a
well-understood rating function. For simColor we want to be able to know
which color they want to switch to, for example.

Adds tests for simColor. Fixes bug in the counting of steps to click the
button then paint.