simChooseColor adds current action to distance.
There was a bug in simChooseColor, similar to a recent bug discovered in
simPaint, where the distance was not accounting for the current action.
For example, movement straight up or down does not usually get the
cursor closer to the button.

To fix this, measure distance from where the cursor will be and add one
for the current action.