Copy states from PICO-8 gamekitty.
I intend to use the gamekitty animation as the starting point for
designing the data structures for describing the state transitions and
animations.

The convert_pico8.py script does a straight conversion from PICO-8 Lua
code to JSON. In the future, I'd like to transform the "s" array to
include additional information like the x-translation and flipx
information.